Move versioning files, update getTemplateWithAssignedUsers
