+++ lib/templates/mutations/createDraftForTemplate.ts
@@ -3,12 +3,12 @@ import {
   isTemplateVersioningEnabled,
   templateRecordInclude,
   templateVersionSelect,
-} from "../internal/versioning";
+} from "../versioning";
 import { authorization } from "@lib/privileges";
 import { AuditLogAccessDeniedDetails, logEvent } from "@lib/auditLogs";
 import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import { TEMPLATE_VERSION_STATUS } from "../internal/types";
-import { parseTemplate } from "../internal/versioning";
+import { parseTemplate } from "../versioning";
 
 export async function createDraftVersionForTemplate(formID: string): Promise<FormRecord | null> {
   if (!(await isTemplateVersioningEnabled())) {
